re-org streaming
